python-tests/requirements.txt
# Drip SDK Test Suite Dependencies

# Drip SDK from PyPI
drip-sdk>=0.1.0

# Testing framework
pytest>=7.0.0
pytest-asyncio>=0.21.0
pytest-json-report>=1.5.0

# Environment configuration
python-dotenv>=1.0.0

# Optional: for better async support
aiohttp>=3.8.0
